[ShellScript] Rework coproc meta scopes
This commit replaces the `meta.function-call` by `meta.coproc` in
coproc statements inspired by `meta.function` in order to clearly
identify parallel running processes.

The scopes `storage.type keyword.declaration` are assigned to `coproc`
because it declares a function which is executed in a separate process
and can be accessed via $COPROC[] system variable or by the defined
name (entity.name.coproc) in the invoking shell script.