[Erlang] Tweak some scopes (#2426)
* [Erlang] Rename variable.other.field to variable.other.member

* [Erlang] Rename entity.name.field to variable.other.member

In Java, field declarations are scoped as variable.other.member as well.

* [Erlang] Rename variables to variable.other.readwrite

The `readwrite` should make clear variables beginning with capital
letters are meant for reading and writing values, while lower case atoms
are constants.

* [Erlang] Use storage.type.record

The #record{field=value} term represents an instantiation of a record,
which was declared by `-record(record, {field=value})` statement. Thus
the #record identifier can be interpreted as storage.type.

This commit uses `storage.type.record` instead of `storage.type.struct`
as this scope name is used for the meta scopes as well.