[HTML] Convert href attribute to old syntax style
While trying to convert the other attribute contexts to the style of
href-attribute, it turned out using `pop:2` doesn't play well in
situations of using contexts in multi-push statements. It certainly
fails for `tag-attributes-generic` whose value is used in several other
places.

In short the reuse-ability of contexts using `pop: 2` is limited.

As the HTML syntax is intended to act as basic syntax for specialized
templating syntaxes etc. all contexts should maintain a maximum of
possible reuse-ability.

Hence converting href-attributes to multi-push style.