[ShellScript] Rework identifier breaks
While  `metachar` denotes the main rule for tokenization, there are
several special rules to pay attention to when handling command names,
command options or variables.

This commit adds variables for each class of identifier to
a) make those differences a bit more clear and
b) improve consistency with regards to when they are applied.

It effects normal and generated rules.