[ShellScript] Rework vardef-value
This commit fixes an issue with scalar values in sequences or mappings
not being highlighted properly.

It therefore restructures how quoted strings may be part of unquoted
strings in scalar values to avoid duplicate meta.string scopes and to
avoid lookaheads for interpolation boundaries.

It is not required to have a single context for each type of quoted
string as they are all supported everywhere. Hence they are grouped in
`strings-quoted`.