[ShellScript] Improve option interpolation
This commit adds `meta.parameter.option.shell` scope to command options
to maintain boundaries if they contain interpolation. It's basically the
same strategy as being used by `meta.string`.

This change is required to avoid duplicated `variable` scopes within
interpolation. It will become important for Makefile.sublime-syntax.

Note: Option values are not scoped as it can't be done consistently for
      unknown external commands. We just don't know, whether an option
      is followed by a value or not.