[ShellScript] Fix `let` builtin arguments
The `let` builtin denotes an expression, which already defines the
allowed expansions. Thus `cmd-args-boilerplate` is not an appropriate
context to include in order to maintain arguments boundaries as it
contains `expansions-and-strings`, which conflicts with expressions.

Regexp pattern expansion is not supported in expressions for instance.

This commit also adds the missing `array-item-access` context to
expressions.