[ShellScript] Fix `test` builtin arguments
The `test` builtin denotes a test expression, which already defines the
supported expansions. Thus `cmd-args-boilerplate` is not an appropriate
context to include, because it contains `expansions-and-strings`, which
conflicts with the supported expansions of the test expressions.

Regexp pattern expansion is not supported on the left side of
comparison operators for instance.

This commit also adds the missing `array-item-access` context to
expressions.