[ShellScript] Improve quotation removal representation
This commit scopes quoted parts of a command name as `variable.function`
instead of `string.quoted` because bash performs so called quotation
removal after variables have been expanded. Thus the former `string` is
no real string but a kind of quoted identifier. Thus it should be scoped
as such. The `meta.string` scope is still applied to identify those
parts.