[Perl] Fix ambiguous function arguments (#2434)
Fixes #2432

Using generic patterns or substitutions such as `/<pattern>/<flags>` as
function arguments without enclosing them into parentheses may be an ambiguous expression as it can't be distinguished from arithmetic
operations in many situations. Things become even more difficult if
generic patterns span multiple lines or contain semi-colons.

This commit splits the list of builtin functions into those which are
known to not support pattern like arguments and those which might do.

These lists might not be perfect. May be tweaked in future, when further
issues about broken highlighting arrive.

User defined functions require arguments to be enclosed in parentheses,
if the first one is a pattern or substitution.