[HTML] Overhaul completions (#2438)
* [HTML] Completions Codestyle

* [HTML] Remove get_completions() method

Why that exists?

* [HTML] Simplify get_tag_to_attributes() function

Don't see a reason why to call extends() to merge two lists which can
be defined as one.

* [HTML] Global get_entity_completions() function

This commit moves the method to build the list of entity completions
out of the class as it has no relation to the object itself. It's a
helper function.

* [HTML] Global get_tag_completions() function

This commit moves the method to build the list of entity completions
out of the class as it has no relation to the object itself. It's a
helper function.

* [HTML] Rework completions

Fixes #2412
Fixes #sublime_text/issues/3476

This commit does a lot in order to

a) improve code style (fix linter errors and use longer variable names)
b) simplify the code a bit
c) fix some of the known issues
d) remove any kind of pre-filter to let ST's AC system decide what to
   suggest. It is no longer useful as `<` is used as tag completion
   trigger.
e) cache completions to reduce computing time on every following call.
   Cold entity completion takes 12ms, while cached entries are returned
   within 0,06ms. Tag completions drop from 1.2ms to 0.2ms.

Note:

1. ST fills its intermediate completion list with everything which looks
   like a valid return value (`list`, `tuple` or `CompletionList`). To
   prevent that, we return `None` if we have nothing to  offer.

2. This commit removes the unittests as they do no longer work
   because of several changes with regards to removing pre-filtering.

* [HTML] Add some benchmarking code

* [HTML] Add missing tags

According to MDN, the following missing tags which are not marked
deprecated or obsolete are added:

- dialog
- hgroup (not html5 standard, but not marked deprecated)
- rb
- slot

* [HTML] Add tag name only completions

If the caret is in front of the closing tag punctuation `>` just
complete the tag name instead of firing any full featured snippet.

1. This helps with `ctrl+shift+w` being active.
   see: https://github.com/sublimehq/sublime_text/issues/3508
2. The normal tag completion would add an additional `>` which is too
   much in this situation.

Note:

1. As KIND_TAG_MARKUP is needed in several situations, all KIND_... are
   moved to globals.
2. The `tag_name_completions` attribute reuses `tag_attributes` to
   retrieve a list of all supported tag names from its keys.

* [HTML] Extend incomplete tag scoping

This commit scopes incomplete tags as `meta.tag.incomplete.html` and
adds the punctuation scopes in order to help completions with
identifying it as candidate.

Also adds incomplete self-closing tags highlighting.

* [HTML] Disable completions in Markdown

HTML completions are not really useful in plain markdown text.
They are disabled except in fenced code blocks.