[Makefile] Improve shebang handling
This commit fixes a regression with enabling shebang highlighting in
Bash breaking Ruby syntax.

It looks like highlighting shebang only at the first line is a bit
tricky with regards to including a syntax into other files.

This commit therefore uses `Shell-Unix-Generic` as intermediate syntax
without shebang patterns which can be included as `scope:source.shell`
as it used to be possible before, while also enabling shebang
highlighting in bash.