[CSS] Replace remaining `keyword.other.unit`
This commit replaces the last remaining `keyword.other.unit` as it was
decided to use `constant.numeric.suffix` for all kinds of units and storage type denoting suffixes in numeric literals.

This commit handles the only situation units appear as standalone token
without a numeric literal in front of them.