[Matlab] Restore special scope for double-percentage
This commit re-adds a pattern to scope comments beginning with %%.

The scope is modified to `comment.line.double-percentage` to comply with
scope naming guidelines more closely. No matter how many `%` chars are
used - it keeps a line comment.

Color Schemes might want to distinguish between `comment.line` and
`comment.block`, which is now possible.

Note:

The `$` is removed for performance reasons. It is not crucial here as
`.*` already implies to match everything in the line.