[CSS] Add at-rule other fallback context
Completions may want to see a `meta.at-rule` and we might not want to
fail on unknown/experimental/new at-rules even though special constants
can't be highlighted then.

This commit therefore adds a general `at-other` context which fulfills
the least syntactical rules to highlight at-rules according to
https://www.w3.org/TR/css-syntax-3/#at-rule

Note:

Any `@<ident> ... [ {} | ; ]` is valid syntax even though the rule might
not (yet) be implemented. A syntax definition is to highlight proper
syntax. It should not be used for linting too much. This is what all the
various LSP plugins or SublimeLinter can do.