[CSS] Fix #id tokens
This commit introduces a `name` variable and uses it instead of
`{{nmchar}}+` in various situations. This is to improve compliance with
CSS tokenization specification.

The `name` is defined at https://www.w3.org/TR/css3-selectors/#lex and
is matched in HASH (after `#`) and IDENT (after `--`).

The main goal is to fix `#id` selector highlighting if the id is of
numeric type.