[Lua] Don't trip on unexpected characters
expression-end used to pop on any otherwise unrecognized character, but
in the context of the expression context being pushed on the stack as
part of a statement, this meant the expression was terminated
unexpectedly early and the following keyword match, such as in `if …
then` would fail to match and also exit early, resulting in an
unfortunate context stack.

The introduced match just matches any unexpected characters and
highlights them as invalid, while also not popping, so the match can
resume after the character.