[CSS] Enforce identifier break
With regards to {{nmchar}} variable \b is not a sufficient pattern to
match identifier or name boundaries.

This commit therefore replaces all trailing \b by a custom pattern.
As using (?!{{nmchar}}) doubles parsing time an optimized alternative is
implemented as {{break}}.

Note: To be able to remove leading \b we'd need to ensure to match all
      tokens in all contexts, which look like names or identifiers.