[CSS] Handle unknown at-rules
Unknown at-rule keywords are to be ignored, while maintaining bracket
and terminator balancing;

see: https://www.w3.org/TR/CSS22/syndata.html#parsing-errors

At-rules with unknown at-keywords. User agents must ignore an invalid
at-keyword together with everything following it, up to the end of the
block that contains the invalid at-keyword, or up to and including the
next semicolon (;), or up to and including the next block ({...}),
whichever comes first.