[CSS] Consistent case insensitive patterns
This commit turns all into case insensitive patterns. Some patterns
already were, others not. According to CSS specs all tokens are to be
matched case insensitive. May not be the best style but doesn't create
parse errors in browsers.

Note: This causes lexing times to increase, so we are where we started.
      The original test file requires 16ms (was 17ms).