[CSS] Revert punctuation.accessor to punctuation.definition
CSS denotes the different kinds of selectors by the opening punctuation.
The concept of accessor would either need to be applied to all those
elements or to nothing. With the usage of stand-alone pseudo classes as
identifier of a page at-rule, it feels more appropriate to keep going
with `punctuation.definition` in all situations.