[CSS] Consolidate identifiers
Any custom CSS identifier may look like `--1dent`, `-name1` or `name2`.
In some situations only the first one is appropriate, in others we might
only match the latter two as the first one is handled separately. See
custom-property for instance.

This commit splits `ident` variable into those two major groups to be
used where needed. `punctuation.definition` is removed from custom
properties for consistency reasons. To keep it there would mean to add
it to any other custom property, which includes quite a number of
patterns to customize, which felt not worth it at this point.

The class #id must consist of a valid custom identifier so `#1` is not
valid.