[HTML] Add dedicated CSS/JS content contexts (#2555)
This commit intends to make it easier for an inherit HTML syntax to
replace the plain `source.js` and `source.css` by inherit definitions.

A syntax such as ASP or JSP needs to inherit CSS and JS in order to
extend them by ASP/JSP tag support. Those extended CSS/JS variants need
to be embedded in HTML-ASP or HTML-JSP then.

The strategy may be compared to what Makefile does with Bash via
Makefile Shell.sublime-syntax.

Unfortunately the `embed:` command doesn't accept variables which could
be used to replace the scope name more easily/consistently. So the
dedicated context is the only way to minimize required code duplication.