[YAML] Don't match controversial plain bools in keys (#2566)
YAML 1.1 made the prominent mistake of accepting many aliases for
boolean values that many parsers adopted. For 1.2, these aliases have
been removed, but for accuracy with the older version and especially
considering how older parsers are still seeing use, we cannot simply
remove matching them.

However, the short names "y" and "on" are fairly common in keys and
having those highlighted as a boolean is annoying, so we only match for
the 1.2 scalars in keys.