[HTML] Exclude prototype from strings by default (#2588)
* [HTML] Exclude prototype from strings by default

This commit moves the `meta_include_prototype` settings from ASP over to
HTML package because it turns out strings to need special handling for
proper `meta.string meta.interpolation source.embedded.xyz.html` in
general.

The reason is the `string` scope which needs to be cleared. As a result
patterns added to prototype may not be injected into string and
attribute value contexts as it would make it impossible to override
those with special patterns.

To handle that ASP used to override a lot of attribute value contexts
of HTML just to add `meta_include_prototype: false`. Any inherited
templating syntax would need to do so, which seems cumbersome.

So it feels more convenient to exclude prototype by default if it is
considered useless or might prevent proper handling of "special cases".