[D] Fix switch case statements
Fixes #2567

This one should have gone into a separate PR, but the root cause is
exactly the same as with issue #2558. Thus the solution is nearly the
same. More than that it makes use of parts of those fixes.

Problem:

The `value-list` content pushes into `value` which matches identifiers
followed by colon as mapping key.