[ShellScript] Fix variable parameter expansion
Addresses one issue of #2635

Bash knows about the special parameter `$#` (number of parameters)
see: https://www.gnu.org/software/bash/manual/bash.html#Special-Parameters

It may be written as parameter expansion `${#}`.

Before this commit the first `#` is always scoped `keyword.operator` as
it is meant to be used with `${#foo}`, which returns the length of
`foo`'s content.

This commit adds a special pattern to match `${#}` as expansion of a
`variable.language`.