[Matlab] Merge built-in and toolbox function contexts
This commit introduces a single context for all built-in functions,
which previously were separated into various contexts while using only a
single regex match each and being referenced only once. The unusual
scoping with 'keyword' or 'storage.type' for built-in functions will be
addressed in a later commit. Correspondingly, another context is used
to merge all the toolbox function contexts.