[Matlab] Improve scoping for operators
This commit adds the corresponding subscopes like 'arithmetic',
'comparison' or 'logical' to operator characters, instead of the unusual
'keyword.operator.symbols' scope, which was used previously.

It also fixes the whitespace around an equal sign '=' to be scoped as
'keyword.operator'.

A few symbols like '>=', '>' were rearranged to ensure that they are
scoped as a single token, so that ligatures can be applied to them.