[Matlab] Improve handling of class declarations
This commit assigns proper scopes to class declarations, including
property, method, event and enumeration block definitions, adds the
missing attribute syntax for event blocks, and fixes class declarations
with multiple superclasses.

For now, the meta scopes are still limited to only the first declaration
lines instead of the whole block. It would be desirable if the meta
scopes would span the whole block (until an 'end' keyword), but that
would require to include almost all of the other contexts for
expressions, function definitions, etc. here as well.