[Matlab] Update list of built-in functions and fix their scopes
This commit updates the built-in function names to those listed in the
most recent MATLAB (R2020b) docs at
https://www.mathworks.com/help/matlab/referencelist.html?type=function
and changes the scopes for the majority of them from `keyword.operator`,
`keyword.other`, `storage.modifier` or unusual scope names like
`keyword.mathematics`, `storage.control` to `support.function.builtin`.
Several hundreds of built-in functions were scoped as `keyword` before,
what made it more difficult to distinguish them from actual control
keywords like `if` or `while`.

In MATLAB, functions with string arguments can be used either in command
syntax form (`funcname param`) or in function syntax form
(`funcname('param')`), see
https://www.mathworks.com/help/matlab/matlab_prog/command-vs-function-syntax.html.

Some essential functions like `tic` or `close all` which usually are
used only in command syntax are scoped as `keyword.other.command` when
it makes sense to emphasize them. Predefined parameter constants for
these commands like `on`, `off` or `all` get `support.constant`, but
only when it is a valid parameter for a specific command.

I've commented out a few commands like `doc` or `whatsnew` which only
really make sense to use in the MATLAB command window (REPL) and not in
files. For now, I've also commented out functions which were removed,
deprecated or are not recommended, because I think it's visually more
appealing not to highlight them, instead of using e.g. the
`invalid.deprecated` scope whose highlighting stands out very much in
many color schemes and can be distracting.

Functions like `pi` or `intmax` which are called without parameters and
return a numeric constant, and functions like `true` or `Inf` which can
be called with or without parameters but are strongly related to
built-in constants still get the `constant.language` scope for
consistency with other syntaxes. Other less essential functions without
parameters like `newline` or `tempdir` which return a (possibly machine
dependent) constant get the `support.constant` scope.

Type cast functions for numeric variables like `double` or `int32` still
get the `storage.type` scope which was used for them before. This scope
is applied on top of `support.function`, because they require an input
parameter and afaik can only be used in function syntax from.