[Matlab] Remove list of toolbox functions
The list of toolbox functions is very outdated, half of the available
toolboxes are not even included and several names are duplicates of
built-in functions. There are already several thousands of built-in
functions without the toolboxes, so I decided to remove the toolbox
functions. Also, if a variable is coincidentally named like a function
from a toolbox, highlighting it as built-in function can be confusing.