[Matlab] Exclude keywords contexts from brackets
Don't include the class_declaration and keywords contexts within
parentheses, brackets or braces. Class definitions or control keywords
are not allowed there and it makes no sense to highlight them.

It might be useful to apply the `invalid.illegal` scope if such keywords
appear as function arguments or variables in parens/brackets/braces,
like e.g. the Python syntax does, but for now it's sufficient to just
not highlight them.