[Matlab] Rename and reorder context inclusions
This commit renames the `allofem` context to `statements` and adds some
more contexts to it. This context includes most of the Matlab language
features, except for declaration and control keywords. It is currently
used within parens/brackets/braces, but could also be used in a later
commit to add proper meta scopes to declaration and control blocks.