[Matlab] Remove invalid.illegal for != characters
This removes the `invalid.illegal` scope for the `!=` character
combination (the "not equal" operator in Matlab is `~=`). It was poorly
implemented because it included the surrounding whitespace, and neither
is such kind of highlighting for random invalid character combinations
used in other syntaxes.

More important, the `!` character is a valid operator (not implemented
in this syntax yet) to denote a operation system command to be executed
in a shell, even though it's unlikely (or impossible?) that it will be
followed by a `=` character.