[Matlab] Remove check for constants override
This commit removes the check for variable assignments that would
override a built-in constants/variables like `ans` or `Inf`. It was
poorly implemented, because it assigned the `meta.inappropriate.matlab`
not only to the identifier, but also to the surrounding whitespace and
`=` character. I guess it was supposed to remove highlighting for
variable names that would override those built-in constants, but I think
such kind of highlighting can actually be helpful, so that one easily
recognizes when choosing inappropriate variable names. And in case of
`i` and `j` as variable names, this highlighting as built-in constant
(imaginary unit) was properly removed in a previous commit.

Furthermore it only worked in certain cases where a semicolon or
whitespace preceded the variable assignment, and the semicolon case even
became obsolete with an earlier commit which added proper scoping for
the semicolon as statement terminator.