[Matlab] Add meta scopes to blocks
This commit adds the corresponding meta scopes to the following
constructs:

- function declaration blocks
- argument validation blocks in function declarations
- class declaration blocks
- properties, methods, events, enumeration blocks in class declarations
- if, elseif, else statements
- switch, case, otherwise statements
- for loops
- parfor loops
- while loops
- try, catch blocks
- spmd blocks

Previously, those meta scopes were only used for the first line in
function declarations and class declarations.

Nested functions or combinations of loops or conditional statements get
stacked meta scopes as expected.

Additionally, support for the `Repeating` attribute in argument
validation blocks was added, and some keyword scopes for the mentioned
constructs were slightly adjusted to be more specific.