[Matlab] Fix function output scope if no whitespace before '='
This commit fixes a special case with wrong scoping for a single
function output argument if there is no whitespace between the argument
name and the equal sign. This issue has been present in this syntax even
before, and I missed it when updating the code for function declarations
in a previous commit.