[Matlab] Fix numbers with trailing decimal point
Numbers that end with a decimal point or a decimal point before the
exponent part or imaginary unit are valid numbers in Matlab.
For example:
  1.   1.e1   1.i

These were not fully recognized as a number, which will be fixed by this
commit now. I'm not sure why I missed these cases earlier.

There is a feature in this syntax to highlight invalid variable
identifiers with `invalid.illegal`, and in particular this applies to
"invalid" identifiers which would start with a number, like `123foo`.
For this to work, the regex for numbers require a word boundary at the
end. I'm not sure whether this is a good approach, because the error
message and error highlighting in Matlab for such a case don't specify
the number part as invalid, but only the subsequent characters. Maybe
this check for valid identifiers should be adjusted to simplify the
number regex, or it should be removed altogether.