[Matlab] Improve detection of transpose operator
This commit makes some improvements for the detection of the transpose
operator `.'` and complex conjugate transpose operator `'`. For the
transpose operator, preceding whitespace is often allowed.
It's also allowed to use the transpose and complex conjugate transpose
operators after scalar numbers (even though this wouldn't really make
sense and is probably never used), strings, and many built-in functions
even when they are called without parentheses (e.g. `true`) and return
an appropriate data type.

I realized that I introduced a regression for the transpose operator
if it occours after built-in functions in command syntax, when I merged
the `transpose` and `transpose_post_parens` contexts. I will fix this in
in a later commit. But in general, the approach to check for a possible
transpose operator after certain contexts used in this Matlab syntax
isn't optimal and could break highlighting if the ' character is used as
transpose operator instead of a single quoted string, but not detected
correctly, as in the examples which are fixed with this commit.