[Matlab] Fix transpose after built-in functions in command syntax
This fixes an issue that was introduced in an earlier commit when the
contexts `transpose` and `transpose_post_parens` were merged. The former
of those was originally included before all of the built-in functions
and therefore took precedence if the transpose operator was used after a
built-in function when it was called in command syntax form (i.e.
without parentheses). In practice this should only happen when the name
of a built-in function is overridden by a variable identifier. So this
commit reintroduces a rule for this special case, and ensures that
whitespace is treated correctly when allowed before the transpose
operator.