[Matlab] Adjust highlighting for overrides of built-in functions
This commit adds a rule to scope built-in function names as ordinary
variable identifiers if they are overridden and used as a variable.
Note that this is different from a rule that was removed in an earlier
commit and did a similar thing only for a subset of built-in variables
and constants. As expained in an earlier commit message, I think in that
case it's desirable that the original scoping with `variable.language`
or `constant.language` is preserved.