[Matlab] Use storage.type only if no function syntax
This commit adjusts the scope `storage.type` to only be used when there
is no opening parenthesis following the type. This will usually be the
case in function argument validation blocks and properties blocks.
Also, add the other fundamental data types besides numeric types for
this scope.

If followed by a parenthesis, use the scope `support.function.builtin`
instead, like for other built-in functions.

I think this is the best solution to use the corresponding scope for
these data type functions based on how they appear in the code.