[Matlab] Ensure to pop at eol in a few contexts
Class and function definitions use a few anonymous contexts for
arguments and attributes in parens/brackets. This commit ensures to pop
at the end of line, if no closing parenthesis/bracket is found (unless
there are line continuation dots). This is required to highlight the
rest of the file correctly, even if there is a incomplete line in a
class or function definition.