[Scala] Scope `extends` and `with` as storage.modifier
The reason for this change is:

It was agreed in a former discussion on discord to only scope class
declaration keywords such as `class`, `object`, etc. as
`keyword.declaration`, while continue using `storage.modifier` for
keywords which modify the declared class entity.

The keywords `extends ... with` point to the base class being extended
by the declared class and thus those modify the class declaration.