[Haskell] Fix comments
Fixes #2670

According to https://www.haskell.org/onlinereport/haskell2010/haskellch10.html
a comment starts with at least 2 dashes followed by `any` character but
symbols with some exceptions. The relevant rules are:

  comment   -> -- {-} [ any⟨symbol⟩ {any} ] newline
  symbol    -> ascSymbol | uniSymbol⟨special|_|"|'⟩
  special   -> ( | ) | , | ; | [ | ] | ` | { | }
  ascSymbol -> ! | # | $ | % | & | ⋆ | + | . | / | < | = | > | ? | @ | \
             | ^ | | | - | ~ | :
  uniSymbol -> any Unicode symbol or punctuation

With:

  any       -> graphic | space | tab
  graphic   -> small | large | symbol | digit | special | " | '

we can say:

  comment   -> -- {-} [
               space | tab | small | large | digit | special | " | '
               {any} ] newline

This simplified pattern is implemented in this commit.

Additionally `(--)` and `(---)` are excluded from infix operators.