[Haskell] Fix literal chars vs. operators
This commit removes invalid highlighting from character literals as this
is not the way Haskell compiler works, which caused syntax highlighting
to break in various situations when leading `'` is used as operator.

Several test cases for char literals are added as well as those to
illustrate how `'` applies as operator or part of an identifier.