[CSS] Explicit constant patterns
This commit implements `numeric-constants` context by using explicit
patterns rather than including the various number type contexts.

The rules are copied from `calc-function-arguments-content`.

The reason for this change is a 16kb smaller *.rcache file and a
somewhat better performance.

Note: As a result some of the special `...-constants` contexts is no
      longer referenced. They keep untouched for future use, if needed.