[XML] Provide dedicated syntax for XML Stylesheets and Transformation
XSL is basically a modern form of DTD.

It is XML syntax and just uses namespaced tags such as `<xs:type>` or
`<xsl:match>`.

This commit adds a dedicated XSL.sublime-syntax which is basically an
alias of XML.sublime-syntax in order to provide a separate top-level
scope for such kinds of files. The result is

XML: text.xml
DTD: text.xml.dtd
XSL: text.xml.xsl

The benefits are:

1. XSL related completions and snippets can be restricted to that file
   type.
2. LSP plugin can send the correct language id "xsl" to a language
   server without special treatment.
3. Dedicated icons may be assigned without plugin interaction or
   dynamically created alias syntaxes.