[Scala] Tweak declaration keyword scopes (#2666)
* [Scala] Scope `package` keyword.declaration.namespace

* [Scala] Scope `import` keyword.declaration.import

* [Scala] Rename `storage.type.function` to `keyword.declaration.function`

* [Scala] Rename `storage.type.class` to `keyword.declaration.class`

* [Scala] Scope `extends` and `with` as storage.modifier

The reason for this change is:

It was agreed in a former discussion on discord to only scope class
declaration keywords such as `class`, `object`, etc. as
`keyword.declaration`, while continue using `storage.modifier` for
keywords which modify the declared class entity.

The keywords `extends ... with` point to the base class being extended
by the declared class and thus those modify the class declaration.

* [Scala] Scope `case` keyword.declaration.other

As `case` in `case class` is scoped keyword.declaration.class it seems
logical to scope it `keyword.declaration.other` in all other cases.