[Matlab] Tweak scopes for end keyword
This commit adjusts some scopes for the `end` keyword to match the
sub-scope of the keyword which was used to begin a specific statement or
block. This ensures that a corresponding `end` has the same highlighting
color as the opening keyword (`function`, `if`, `while`, ...).

Additionally, two small adjustments for the `methods` and `spmd` blocks
were made to use a `keyword.context` scope, which seems more fitting for
these.