[XML] Rename entity-parameter context
This commit is to make clear `#PCDATA` and `%name;` tags belong to
doctype definitions rather than normal xml. That's why the defining
context is renamed to `dtd-constants`.

The renamed context is rearranged in the DTD PROTOTYPES section because
`dtd-common-names` also contains `%name;` patterns and it feels natural
from naming perspective.

Note: There are no known usages of that context in 3rd-party packages.