[Markdown] Fix keymap contexts
This commit removes the restriction of `preceeding_text` to need to be
nonwhite space in order to enable bold/italic or backtick auto-pairing.

That's a quite uncommon restriction as a user might expect to be able
to start a new bold word preceded by space.

Example (didn't work before):

  This is |       =>        This is *|*

On the other hand auto-pairing might not be desired if escape backtick
is preceeded:

  This is \|      =>        This is \*|

but should if arbitary escaped backslashes are:

  This is \\|     =>        This is \\*|*
  This is \\\\|   =>        This is \\\\*|*