[XML] Enhance inheritability (#2699)
* [XML] Replace pop: true by pop: 1

This commit replaces all `pop: true` to clearly reflect XML using
sublime-syntax version 2 as done for all other version 2 syntaxes.

* [XML] Improve highlighting with missing attributes

This commit relaxes the restriction for value assignments in xml attributes. A `=` sign is now accepted even if the attribute name is not
present.

This change helps with incomplete codes and enables correct highlighting
in case the attribute name being an injected token.

Example:

    <tag {{jinja_attribute_template}}="value"/>

* [XML] Add meta.string and exclude prototypes

This commit ...

1. adds `meta.string.xml` to all strings
2. excludes prototype from all strings

The goal is to help template languages which inherit from XML to
properly handle string interpolation. The strategy is very much the same
as already implemented in HTML.

* [XML] Introduce named tag content contexts

This commit turns all anonymous contexts into named contexts in order to
help template languages which inherit XML to modify single aspects
easier.

* [XML] Rename entity-parameter context

This commit is to make clear `#PCDATA` and `%name;` tags belong to
doctype definitions rather than normal xml. That's why the defining
context is renamed to `dtd-constants`.

The renamed context is rearranged in the DTD PROTOTYPES section because
`dtd-common-names` also contains `%name;` patterns and it feels natural
from naming perspective.

Note: There are no known usages of that context in 3rd-party packages.