[CSS] Add smart_typing setting
This commit adds support for `smart_typing` setting to extend scope of
`auto_match_enabled` to related but somewhat smarter typing features.

The intention is to satisfy both the lazy and convenient developer and
the purist who needs control over every single key stroke.

Due to several requests some input helpers were introduced during last
dev cycle, which intend to help keeping focus on content while helping
with some nasty repetitive tasks like adding space after colons and
move on to next line when hitting enter in front of semi-colons.

Now that those smart behaviors exist, of course, some people complain
about it.

With this commit following features are enabled only 
if `smart_typing` is `true`:

1. Automatically add space after colons.
2. Hitting enter in front of semicolon moves to next line.
3. Hitting space in between brackets adds space to left and right
   position of caret.

Note:

The `smart_typing` setting is NOT added to CSS.sublime-settings as it is
intended to be used for related features in all syntax packages, just
like `auto_match_enabled`.

Adding it to CSS would override user defined global settings in
Preferences.sublime-settings.

As a result all those smart typing features are disabled by default in case this setting is NOT added to Preferences.sublime-settings of the
Default.sublime-package or if it is decided to disable such feature by
default.